feat: implement key generation and migration for vaults
- Introduced `CreateVaultKeygenActionProvider` and `MigrateVaultKeygenActionProvider` to handle key generation and migration processes for vaults.
- Added `ReshareVaultKeygenActionProvider` to facilitate the reshare functionality for existing vaults.
- Refactored key generation logic to utilize DKLS and Schnorr algorithms, enhancing security and modularity.
- Implemented new hooks for managing key generation state and actions, improving the overall structure and usability of the vault key management system.
- Removed obsolete files and utilities related to previous key generation methods, streamlining the codebase.
