+++ clients/desktop/src/vault/keygen/reshare/ReshareVaultLegacyKeygenActionProvider.tsx
@@ -1,0 +1,111 @@
+import { keygenSteps } from '@core/mpc/keygen/KeygenStep'
+import { MpcLib } from '@core/mpc/mpcLib'
+import { KeygenAction } from '@core/ui/mpc/keygen/state/keygenAction'
+import {
+  KeygenVault,
+  useKeygenVault,
+  useKeygenVaultName,
+} from '@core/ui/mpc/keygen/state/keygenVault'
+import { useCurrentHexEncryptionKey } from '@core/ui/mpc/state/currentHexEncryptionKey'
+import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
+import { useMpcServerUrl } from '@core/ui/mpc/state/mpcServerUrl'
+import { useMpcSessionId } from '@core/ui/mpc/state/mpcSession'
+import { useVaults } from '@core/ui/vault/state/vaults'
+import { Vault } from '@core/ui/vault/Vault'
+import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
+import { getLastItemOrder } from '@lib/utils/order/getLastItemOrder'
+import { EventsOff, EventsOn } from '@wailsapp/runtime'
+import { useCallback } from 'react'
+
+import { storage } from '../../../../wailsjs/go/models'
+import { Reshare } from '../../../../wailsjs/go/tss/TssService'
+import { fromStorageVault, toStorageVault } from '../../utils/storageVault'
+
+export const ReshareVaultLegacyKeygenActionProvider = () => {
+  const serverUrl = useMpcServerUrl()
+  const encryptionKeyHex = useCurrentHexEncryptionKey()
+  const sessionId = useMpcSessionId()
+  const keygenVault = useKeygenVault()
+  const vaultName = useKeygenVaultName()
+  const localPartyId = useMpcLocalPartyId()
+  const vaults = useVaults()
+
+  const vaultOrders = vaults.map(vault => vault.order)
+
+  const keygenAction: KeygenAction = useCallback(
+    async ({ onStepChange }) => {
+      keygenSteps.forEach(step => {
+        EventsOn(step, () => onStepChange(step))
+      })
+
+      const sharedFinalVaultFields = {
+        localPartyId,
+        libType: 'GG20' as MpcLib,
+        isBackedUp: false,
+      }
+
+      try {
+        return matchRecordUnion<KeygenVault, Promise<Vault>>(keygenVault, {
+          existingVault: async existingVault => {
+            const storageVault = await Reshare(
+              toStorageVault(existingVault),
+              sessionId,
+              encryptionKeyHex,
+              serverUrl
+            )
+
+            return {
+              ...existingVault,
+              ...fromStorageVault(storageVault),
+              ...sharedFinalVaultFields,
+            }
+          },
+          newReshareVault: async ({
+            oldResharePrefix,
+            oldParties,
+            publicKeyEcdsa,
+            hexChainCode,
+          }) => {
+            const storageVault = await Reshare(
+              {
+                public_key_ecdsa: publicKeyEcdsa,
+                signers: oldParties,
+                reshare_prefix: oldResharePrefix,
+                hex_chain_code: hexChainCode,
+                local_party_id: localPartyId,
+                name: vaultName,
+              } as storage.Vault,
+              sessionId,
+              encryptionKeyHex,
+              serverUrl
+            )
+
+            return {
+              ...fromStorageVault(storageVault),
+              ...sharedFinalVaultFields,
+              order: getLastItemOrder(vaultOrders),
+            }
+          },
+          newVault: () => {
+            throw new Error('new vault is not expected for Reshare')
+          },
+        })
+      } finally {
+        keygenSteps.forEach(step => {
+          EventsOff(step)
+        })
+      }
+    },
+    [
+      encryptionKeyHex,
+      keygenVault,
+      localPartyId,
+      serverUrl,
+      sessionId,
+      vaultName,
+      vaultOrders,
+    ]
+  )
+
+  return keygenAction
+}
