--- clients/desktop/src/vault/keygen/shared/mutations/KeygenResolver.ts
@@ -1,10 +1,0 @@
-import { KeygenStep } from '@core/mpc/keygen/KeygenStep'
-import { Vault } from '@core/ui/vault/Vault'
-
-type KeygenResolverInput = {
-  onStepChange: (step: KeygenStep) => void
-}
-
-export type KeygenResolver = ({
-  onStepChange,
-}: KeygenResolverInput) => Promise<Vault>
