+++ core/ui/mpc/keygen/reshare/ReshareVaultKeygenActionProvider.tsx
@@ -1,0 +1,153 @@
+import { DKLS } from '@core/mpc/dkls/dkls'
+import {
+  setKeygenComplete,
+  waitForKeygenComplete,
+} from '@core/mpc/keygenComplete'
+import { MpcLib } from '@core/mpc/mpcLib'
+import { Schnorr } from '@core/mpc/schnorr/schnorrKeygen'
+import { useCurrentHexEncryptionKey } from '@core/ui/mpc/state/currentHexEncryptionKey'
+import { useIsInitiatingDevice } from '@core/ui/mpc/state/isInitiatingDevice'
+import { useMpcLocalPartyId } from '@core/ui/mpc/state/mpcLocalPartyId'
+import { useMpcPeers } from '@core/ui/mpc/state/mpcPeers'
+import { useMpcServerUrl } from '@core/ui/mpc/state/mpcServerUrl'
+import { useMpcSessionId } from '@core/ui/mpc/state/mpcSession'
+import { useVaults } from '@core/ui/vault/state/vaults'
+import { getLastItemOrder } from '@lib/utils/order/getLastItemOrder'
+import { useCallback } from 'react'
+
+import { KeygenAction } from '../state/keygenAction'
+import {
+  assertKeygenReshareFields,
+  useKeygenVault,
+  useKeygenVaultName,
+} from '../state/keygenVault'
+
+export const ReshareVaultKeygenActionProvider = () => {
+  const serverUrl = useMpcServerUrl()
+  const encryptionKeyHex = useCurrentHexEncryptionKey()
+  const sessionId = useMpcSessionId()
+  const vaultName = useKeygenVaultName()
+  const localPartyId = useMpcLocalPartyId()
+  const peers = useMpcPeers()
+  const isInitiatingDevice = useIsInitiatingDevice()
+  const vaults = useVaults()
+  const keygenVault = useKeygenVault()
+
+  const vaultOrders = vaults.map(vault => vault.order)
+
+  const keygenAction: KeygenAction = useCallback(
+    async ({ onStepChange }) => {
+      onStepChange('ecdsa')
+
+      const signers = [localPartyId, ...peers]
+
+      const sharedFinalVaultFields = {
+        signers,
+        localPartyId,
+        libType: 'DKLS' as MpcLib,
+        isBackedUp: false,
+      }
+
+      const { oldParties } = assertKeygenReshareFields(keygenVault)
+
+      const oldCommittee = oldParties.filter(party => signers.includes(party))
+
+      const dklsKeygen = new DKLS(
+        'reshare',
+        isInitiatingDevice,
+        serverUrl,
+        sessionId,
+        localPartyId,
+        signers,
+        oldCommittee,
+        encryptionKeyHex
+      )
+
+      const oldEcdsaKeyshare =
+        'existingVault' in keygenVault
+          ? keygenVault.existingVault.keyShares.ecdsa
+          : undefined
+
+      const dklsResult =
+        await dklsKeygen.startReshareWithRetry(oldEcdsaKeyshare)
+
+      onStepChange('eddsa')
+
+      const schnorrKeygen = new Schnorr(
+        'reshare',
+        isInitiatingDevice,
+        serverUrl,
+        sessionId,
+        localPartyId,
+        signers,
+        oldCommittee,
+        encryptionKeyHex,
+        new Uint8Array(0)
+      )
+
+      const oldEddsaKeyshare =
+        'existingVault' in keygenVault
+          ? keygenVault.existingVault.keyShares.eddsa
+          : undefined
+
+      const schnorrResult =
+        await schnorrKeygen.startReshareWithRetry(oldEddsaKeyshare)
+
+      const publicKeys = {
+        ecdsa: dklsResult.publicKey,
+        eddsa: schnorrResult.publicKey,
+      }
+
+      const keyShares = {
+        ecdsa: dklsResult.keyshare,
+        eddsa: schnorrResult.keyshare,
+      }
+
+      const newVaultFields = {
+        publicKeys,
+        keyShares,
+        hexChainCode: dklsResult.chaincode,
+        ...sharedFinalVaultFields,
+      }
+
+      const vault =
+        'existingVault' in keygenVault
+          ? {
+              ...keygenVault.existingVault,
+              ...newVaultFields,
+            }
+          : {
+              ...newVaultFields,
+              name: vaultName,
+              order: getLastItemOrder(vaultOrders),
+            }
+
+      await setKeygenComplete({
+        serverURL: serverUrl,
+        sessionId: sessionId,
+        localPartyId,
+      })
+
+      await waitForKeygenComplete({
+        serverURL: serverUrl,
+        sessionId: sessionId,
+        peers,
+      })
+
+      return vault
+    },
+    [
+      encryptionKeyHex,
+      isInitiatingDevice,
+      keygenVault,
+      localPartyId,
+      peers,
+      serverUrl,
+      sessionId,
+      vaultName,
+      vaultOrders,
+    ]
+  )
+
+  return keygenAction
+}
